Dead code cleanup: remove unused files, extract format_money, fix auth
Deleted (10 files):
- HttpBasicAuthenticable concern (never included)
- movement_demo route + orphan test (no controller existed)
- PWA views (routes commented out)
- content_security_policy.rb (entirely commented out)
- react_server_rendering.rb (referenced non-existent asset)
- blog.rake (duplicate of blog_posts.rake)
- convert_images.rake (target directory doesn't exist)
- ScrollbarStyles.jsx (never imported)
- DebugPanel.jsx (never imported)

Fixes:
- BriefingsController now inherits Api::V1::BaseController (was bypassing API auth)
- Extracted format_money into MoneyFormattable concern (was duplicated 3x)

109 tests, 0 failures. Rubocop clean.
