--- app/views/pwa/manifest.json.erb
@@ -1,22 +1,0 @@
-{
-  "name": "Austn",
-  "icons": [
-    {
-      "src": "/icon.png",
-      "type": "image/png",
-      "sizes": "512x512"
-    },
-    {
-      "src": "/icon.png",
-      "type": "image/png",
-      "sizes": "512x512",
-      "purpose": "maskable"
-    }
-  ],
-  "start_url": "/",
-  "display": "standalone",
-  "scope": "/",
-  "description": "Austn.",
-  "theme_color": "red",
-  "background_color": "red"
-}
