--- config/initializers/react_server_rendering.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-# To render React components in production, precompile the server rendering manifest:
-Rails.application.config.assets.precompile += [ "server_rendering.js" ]
